+++ src/constants.ts
@@ -1,0 +1,19 @@
+export const MODULE_NAME = "fast-flip";
+
+export namespace LOCALIZATION {
+    export const MIRROR_HORIZONTAL_HOTKEY = `${MODULE_NAME}.mirror-horizontal-hotkey`;
+    export const MIRROR_HORIZONTAL_HINT = `${MODULE_NAME}.mirror-horizontal-hint`;
+    export const MIRROR_HORIZONTAL_BUTTON = `${MODULE_NAME}.mirror-horizontal-button`;
+    export const MIRROR_VERTICAL_HOTKEY = `${MODULE_NAME}.mirror-vertical-hotkey`;
+    export const MIRROR_VERTICAL_HINT = `${MODULE_NAME}.mirror-vertical-hint`;
+    export const MIRROR_VERTICAL_BUTTON = `${MODULE_NAME}.mirror-vertical-button`;
+    export const TOGGLE_AFK_HOTKEY = `${MODULE_NAME}.toggle-afk-hotkey`;
+    export const TOGGLE_AFK_HINT = `${MODULE_NAME}.toggle-afk-hint`;
+    export const TOGGLE_AFK_BUTTON = `${MODULE_NAME}.toggle-afk-button`;
+    export const SHOW_MIRROR_BUTTONS = `${MODULE_NAME}.show-mirror-buttons`;
+    export const SHOW_MIRROR_BUTTONS_HINT = `${MODULE_NAME}.show-mirror-buttons-hint`;
+    export const SHOW_TOGGLE_AFK_BUTTON = `${MODULE_NAME}.show-toggle-afk-button`;
+    export const SHOW_TOGGLE_AFK_HINT = `${MODULE_NAME}.show-toggle-afk-hint`;
+    export const AFK_OVERLAY_ICON_PATH = `${MODULE_NAME}.afk-overlay-icon-path`;
+    export const AFK_OVERLAY_ICON_PATH_HINT = `${MODULE_NAME}.afk-overlay-icon-path-hint`;
+};