+++ src/helpers.ts
@@ -1,0 +1,5 @@
+import { MODULE_NAME } from "./constants";
+
+export function getIcon(icon: string): string {
+    return `modules/${MODULE_NAME}/icons/${icon}.svg`;
+}