+++ src/module.ts
@@ -1,39 +1,78 @@
-const MODULE_NAME = "fast-flip";
-const MIRROR_HORIZONTAL_HOT_KEY = `${MODULE_NAME}.mirror-horizontal-hotkey`;
-const MIRROR_VERTICAL_HOT_KEY = `${MODULE_NAME}.mirror-vertical-hotkey`;
-
-const enum TileMirror {
-    HORIZONTAL = "tileMirrorHorizontal",
-    VERTICAL = "tileMirrorVertical"
-}
-
-interface TileFlags {
-    [TileMirror.VERTICAL]?: boolean;
-    [TileMirror.HORIZONTAL]?: boolean;
-}
+import { LOCALIZATION, MODULE_NAME } from "./constants";
+import { TokenManager, TokenMirror } from "./managers/TokenManager";
+import { TileManager, TileMirror } from "./managers/TileManager";
+import { TileHUDManager, TokenHUDManager } from "./managers/HUDManager";
+import { Settings } from "./Settings";
 
 Hooks.once("init", () => {
     if (game instanceof Game) {
-        (game as any).keybindings.register(MODULE_NAME, "horizontalFlip", {
-            name: MIRROR_HORIZONTAL_HOT_KEY,
-            hint: "Horizontally mirrors the selected tile or token",
+        new FastFlipModule(game);
+    }
+});
+
+class FastFlipModule {
+    readonly #tokenHUDManager: TokenHUDManager;
+    readonly #tileHUDManager: TileHUDManager;
+    readonly #tokenManager: TokenManager;
+    readonly #tileManager: TileManager;
+    readonly #game: Game;
+
+    constructor(game: Game) {
+        this.#game = game;
+        const settings = new Settings(game);
+        this.#tokenHUDManager = new TokenHUDManager(game);
+        this.#tileHUDManager = new TileHUDManager(game);
+        this.#tokenManager = new TokenManager(game, settings);
+        this.#tileManager = new TileManager(game);
+
+        this.#registerKeybindings();
+        this.#registerHUDButtons(settings);
+    }
+
+
+    #registerKeybindings() {
+        this.#game.keybindings.register(MODULE_NAME, "horizontalFlip", {
+            name: LOCALIZATION.MIRROR_HORIZONTAL_HOTKEY,
+            hint: this.#game.i18n.localize(LOCALIZATION.MIRROR_HORIZONTAL_HINT),
             editable: [
                 { key: "KeyF" },
             ],
-            onDown: handleHorizontalMirror,
+            onDown: async () => {
+                await this.#tokenManager.mirrorSelected(TokenMirror.HORIZONTAL);
+                await this.#tileManager.mirrorSelectedTiles(TileMirror.HORIZONTAL);
+            },
+            // TODO: Fix this once V9 bindings are out.
             precedence: (CONST as any).KEYBINDING_PRECEDENCE.NORMAL,
             restrictied: false,
             reservedModifiers: [],
             repeat: false,
         });
 
-        (game as any).keybindings.register(MODULE_NAME, "verticalFlip", {
-            name: MIRROR_VERTICAL_HOT_KEY,
-            hint: "Vertically mirrors the selected tile or token",
+        this.#game.keybindings.register(MODULE_NAME, "verticalFlip", {
+            name: LOCALIZATION.MIRROR_VERTICAL_HOTKEY,
+            hint: this.#game.i18n.localize(LOCALIZATION.MIRROR_VERTICAL_HINT),
             editable: [
-                { key: "KeyF", modifiers: ["Shift"] },
+                { key: "KeyF", modifiers: ["SHIFT"] },
             ],
-            onDown: handleVerticalMirror,
+            onDown: async () => {
+                await this.#tokenManager.mirrorSelected(TokenMirror.VERTICAL);
+                await this.#tileManager.mirrorSelectedTiles(TileMirror.VERTICAL);
+            },
+            // TODO: Fix this once V9 bindings are out.
+            precedence: (CONST as any).KEYBINDING_PRECEDENCE.NORMAL,
+            restrictied: false,
+            reservedModifiers: [],
+            repeat: false,
+        });
+
+        this.#game.keybindings.register(MODULE_NAME, LOCALIZATION.TOGGLE_AFK_HOTKEY, {
+            name: LOCALIZATION.TOGGLE_AFK_HOTKEY,
+            hint: this.#game.i18n.localize(LOCALIZATION.TOGGLE_AFK_HINT),
+            editable: [
+                { key: "KeyK", modifiers: ["SHIFT"] },
+            ],
+            onDown: async () => await this.#tokenManager.toggleAFK(),
+            // TODO: Fix this once V9 bindings are out.
             precedence: (CONST as any).KEYBINDING_PRECEDENCE.NORMAL,
             restrictied: false,
             reservedModifiers: [],
@@ -41,78 +80,43 @@ Hooks.once("init", () => {
         });
     }
 
-});
+    #registerHUDButtons(settings: Settings) {
+        this.#tokenHUDManager.registerButton(`${MODULE_NAME}.mirror-horizontal`, {
+            side: "left",
+            title: LOCALIZATION.MIRROR_HORIZONTAL_BUTTON,
+            icon: "mirror-horizontal",
+            onClick: async () => await this.#tokenManager.mirrorSelected(TokenMirror.HORIZONTAL),
+            shouldShow: () => settings.showMirrorButtons,
+        });
 
-Hooks.on("updateTile", (_: unknown, update: foundry.data.TileData) => {
-    if (game instanceof Game && game.canvas?.ready && update._id) {
-        const tile = game.canvas.background?.get(update._id)
-            ?? game.canvas.foreground?.get(update._id);
+        this.#tokenHUDManager.registerButton(`${MODULE_NAME}.mirror-vertical`, {
+            side: "left",
+            title: LOCALIZATION.MIRROR_VERTICAL_BUTTON,
+            icon: "mirror-vertical",
+            onClick: async () => await this.#tokenManager.mirrorSelected(TokenMirror.VERTICAL),
+            shouldShow: () => settings.showMirrorButtons,
+        });
 
-        if (tile) {
-            updateTileOrientation(tile);
-        }
-    }
-});
+        this.#tokenHUDManager.registerButton(`${MODULE_NAME}.toggle-afk`, {
+            side: "right",
+            title: LOCALIZATION.TOGGLE_AFK_BUTTON,
+            icon: "toggle-afk",
+            onClick: async () => await this.#tokenManager.toggleAFK(),
+            shouldShow: () => settings.showToggleAFKButton,
+        });
 
-Hooks.on("canvasReady", () => {
-    if (game instanceof Game && game.canvas?.ready) {
-        const tiles = [
-            ...game.canvas.background?.tiles ?? [],
-            ...game.canvas.foreground?.tiles ?? []
-        ];
+        this.#tileHUDManager.registerButton(`${MODULE_NAME}.mirror-horizontal`, {
+            side: "left",
+            title: LOCALIZATION.MIRROR_HORIZONTAL_BUTTON,
+            icon: "mirror-horizontal",
+            onClick: async () => await this.#tileManager.mirrorSelectedTiles(TileMirror.HORIZONTAL),
+        });
 
-        for (const tile of tiles) {
-            updateTileOrientation(tile);
-        }
-    }
-});
-
-async function handleHorizontalMirror() {
-    if (game instanceof Game && game.canvas?.ready) {
-        const controlledTokens = game.canvas.tokens?.controlled ?? [];
-        for (const token of controlledTokens) {
-            await token.document.update({ "mirrorX": !token.data.mirrorX });
-        }
-
-        const controlledTiles = [
-            ...game.canvas.background?.controlled ?? [],
-            ...game.canvas.foreground?.controlled ?? []
-        ];
-        for (const tile of controlledTiles) {
-            const previousState = tile.document.getFlag(MODULE_NAME, TileMirror.HORIZONTAL);
-            await tile.document.setFlag(MODULE_NAME, TileMirror.HORIZONTAL, !previousState);
-        }
-    }
-}
-
-async function handleVerticalMirror() {
-    if (game instanceof Game && game.canvas?.ready) {
-        const controlledTokens = game.canvas.tokens?.controlled ?? [];
-        for (const token of controlledTokens) {
-            await token.document.update({ "mirrorY": !token.data.mirrorY });
-        }
-
-        const controlledTiles = [
-            ...game.canvas.background?.controlled ?? [],
-            ...game.canvas.foreground?.controlled ?? []
-        ];
-        for (const tile of controlledTiles) {
-            const previousState = tile.document.getFlag(MODULE_NAME, TileMirror.VERTICAL);
-            await tile.document.setFlag(MODULE_NAME, TileMirror.VERTICAL, !previousState);
-        }
-    }
-}
-
-function updateTileOrientation(tile: Tile) {
-    if (tile.texture != null) {
-        const flipHorizontal = tile.document.getFlag(MODULE_NAME, TileMirror.HORIZONTAL);
-        const flipVerical = tile.document.getFlag(MODULE_NAME, TileMirror.VERTICAL);
-
-        const mirrorHorizontal = flipHorizontal ? PIXI.groupD8.MIRROR_HORIZONTAL : 0;
-        const mirrorVertical = flipVerical ? PIXI.groupD8.MIRROR_VERTICAL : 0;
-        const rotate = PIXI.groupD8.add(mirrorHorizontal, mirrorVertical);
-
-        tile.texture.rotate = rotate;
-        tile.refresh();
+        this.#tileHUDManager.registerButton(`${MODULE_NAME}.mirror-vertical`, {
+            side: "left",
+            title: LOCALIZATION.MIRROR_VERTICAL_BUTTON,
+            icon: "mirror-vertical",
+            onClick: async () => await this.#tileManager.mirrorSelectedTiles(TileMirror.VERTICAL),
+        });
     }
 }