+++ Changelog.md
@@ -1,3 +1,10 @@
+# v2.1.0
+* Added a `Shift+K` (default binding) shortcut to mark a token as AFK
+* Added token HUD buttons for mirroring tokens and marking tokens as AFK
+* Added tile HUD buttons for mirroring tiles
+* Added settings to allow players to show/hide the mirror and AFK toggle buttons on their token HUD
+* Restructured the mod internals to allow for more token related features to be added
+
 # v2.0.0
 * Support for Foundry V9
 * Migrate to the new V9 Keybindings API (this is a breaking change that requires V9 minimum)
@@ -14,4 +21,4 @@ Tiles!
 # v1.0.1
 Initial public release of Fast Flip!
 
-* Create a keybind that can be used to horizontally mirror all selected tokens. Works for both the GM and players.
+* Create a keybind that can be used to horizontally mirror all selected tokens. Works for both the GM and players